+++ analytics/lib/user_constants.dart
@@ -1,0 +1,7 @@
+enum Season { spring, summer, autumn, winter, none }
+
+enum PreferredTempUnits { c, f, none }
+
+enum PreferredWeatherTemp { hot, cold, none }
+
+enum PreferredWeatherCond { rain, sun, none }
