+++ analytics/pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   provider: ^6.0.5
   firebase_core: ^2.4.1
+  firebase_analytics: ^10.1.0
 
 dev_dependencies:
   flutter_test:
